requirements.txt
python-telegram-bot==20.7
pymongo==4.3.3
Flask==2.3.3
gunicorn==21.2.0